Refactor Computer Task
