+++ src/lesson_5/com/abstractstasks/transport/types/CivilAirPlane.java
@@ -1,4 +1,4 @@
-package lesson_5.com.abstractstasks.transport.classes;
+package lesson_5.com.abstractstasks.transport.types;
 
 import lesson_5.com.abstractstasks.transport.abstracts.AirTransport;
 import lesson_5.com.abstractstasks.transport.enums.Model;
