+++ src/lesson_5/com/additionaltasks/computer/Computer.java
@@ -1,9 +1,7 @@
 package lesson_5.com.additionaltasks.computer;
 
+import lesson_5.com.additionaltasks.computer.commands.ThrowCoin;
 import lombok.Getter;
-
-import java.util.Random;
-import java.util.Scanner;
 
 @Getter
 public class Computer {
@@ -32,8 +30,8 @@ public class Computer {
         if (isBurned){
             System.out.println("You have no computer anymore!");
         } else if (isWorking){
-        System.out.println("Computer configuration [processor: " + PROCESSOR + ", RAM: " + RAM + ", HDD: "
-                + HDD + ", cycles-to-failure: " + cyclesToFailure + "]");
+            System.out.println("Computer configuration [processor: " + PROCESSOR + ", RAM: " + RAM + ", HDD: "
+                    + HDD + ", cycles-to-failure: " + cyclesToFailure + "]");
         } else {
             System.out.println("Tip: You must turn computer first!");
         }
@@ -41,8 +39,7 @@ public class Computer {
 
     public void turnOn() {
         if (!isWorking && !isBurned) {
-            int coinFlip = new Random().nextInt(10)%2;
-            if (coinFlip == throwCoin()) {
+            if (new ThrowCoin().Execute()) {
                 System.out.println("Computer successfully turned on");
                 isWorking = true;
             } else {
@@ -58,15 +55,14 @@ public class Computer {
 
     public void turnOff() {
         if (isWorking) {
-            int coinFlip = new Random().nextInt(10)%2;
-            if (coinFlip == throwCoin()) {
-            System.out.println("Computer is now turned off");
-            isWorking = false;
-            cyclesToFailure--;
-            System.out.println(cyclesToFailure + " cycles left.");
+            if (new ThrowCoin().Execute()) {
+                System.out.println("Computer is now turned off");
+                isWorking = false;
+                cyclesToFailure--;
+                System.out.println(cyclesToFailure + " cycles left.");
             } else {
-                    burnDown();
-                }
+                burnDown();
+            }
         } else {
             System.out.println("Computer is not working now.");
         }
@@ -77,21 +73,4 @@ public class Computer {
         isBurned = true;
     }
 
-    private int throwCoin (){
-        int numberFromUser = 2;
-        while (numberFromUser != 1 && numberFromUser != 0){
-            System.out.println("Select a number: 1 or 0");
-            numberFromUser = getNumberFromScanner();
-        }
-        return numberFromUser;
-    }
-
-     public int getNumberFromScanner() {
-        Scanner scanner = new Scanner(System.in);
-        int number = -1;
-        try { number = Integer.parseInt(scanner.nextLine().trim());
-        } catch (NumberFormatException e){
-            System.out.println("Wrong number!"); }
-        return number;
-    }
 }
